feat: error handling & minor code refactor

+++ app/api/journal/[id]/route.ts
@@ -1,13 +1,19 @@
+import { NextResponse } from 'next/server';
 import { update } from '@/utils/actions';
 import { analyzeEntry } from '@/utils/ai';
 import { getUserByClerkId } from '@/utils/auth';
 import { prisma } from '@/utils/db';
-import { NextResponse } from 'next/server';
+import { isDynamicServerError } from 'next/dist/client/components/hooks-server-context';
 
 export const PATCH = async (request: Request, { params }: { params: { id: string } }) => {
+  const { content } = await request.json();
+  const user = await getUserByClerkId();
+
+  if (!user) {
+    return NextResponse.json({ message: 'User not found' }, { status: 401 });
+  }
+
   try {
-    const { content } = await request.json();
-    const user = await getUserByClerkId();
     const updatedEntry = await prisma.journalEntry.update({
       where: {
         userId_id: {
@@ -36,14 +42,23 @@ export const PATCH = async (request: Request, { params }: { params: { id: string
 
     return NextResponse.json({ data: { ...updatedEntry, analysis: updated } });
   } catch (error) {
+    if (isDynamicServerError(error)) {
+      throw error;
+    }
+
+    console.error('Error processing PATCH request:', error);
     return NextResponse.json({ message: 'Error processing request' }, { status: 500 });
   }
 };
 
 export const DELETE = async (_request: Request, { params }: { params: { id: string } }) => {
-  try {
-    const user = await getUserByClerkId();
+  const user = await getUserByClerkId();
 
+  if (!user) {
+    return NextResponse.json({ message: 'User not found' }, { status: 401 });
+  }
+
+  try {
     await prisma.journalEntry.delete({
       where: {
         userId_id: {
@@ -57,6 +72,11 @@ export const DELETE = async (_request: Request, { params }: { params: { id: stri
 
     return NextResponse.json({ data: { id: params.id } });
   } catch (error) {
+    if (isDynamicServerError(error)) {
+      throw error;
+    }
+
+    console.error('Error processing DELETE request:', error);
     return NextResponse.json({ message: 'Error processing request' }, { status: 500 });
   }
 };
